Refactor sharing into module.
